Add make_device_builder to test class
The imxrt-usbd driver requires that the user sets the EP0 max packet
size to 64 bytes. To meet the requirement in our test_class example,
we need access to the UsbDeviceBuilder. Ideally, we could access the
builder without duplicating the make_device method.

This commit adds a new method, make_device_builder, for the test
class. It sets the required device properties, then returns the
builder to the caller for any additional configurations.