Don't require UsbBus to be Sync.
If the UsbBus is not Sync, it can still be used to make a UsbDevice,
but that UsbDevice will not be Sync.  This is quite limiting (as it
mostly means that the USB device can't be accessed from interrupt
contexts), but it is entirely safe from the Rust safety perspective.