Separate read_setup() from read()
This change aims to resolve issues where an expected OUT Zero Length
Packet (ZLP) was overwritten by a subsequent SETUP transaction.  The USB
spec requires that devices accept SETUP transactions, this behaviour is
implemented directly in some USB hardware.  To allow the control pipe
implementation to distinguish between OUT data and SETUP data, the
read() method was changed to return an error if SETUP data was present,
and a new read_setup() was added that expects SETUP data not OUT data.